Move reports prop-types to reports folder
