+++ src/pages/Reports/pages/OverviewMap/components/ReportDetails/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
 
-import FMCPropTypes from '~/propTypes';
+import FMCPropTypes from '~/pages/Reports/propTypes';
 import detailWrapped from '~/pages/Map/components/DetailView/detailWrapped';
 import HorizontalRuler from '~/pages/Reports/pages/SubmitReport/components/HorizontalRuler';
