+++ src/pages/Reports/pages/OverviewMap/components/WebglMap.js
@@ -7,7 +7,7 @@ import config from '~/pages/Reports/config';
 import logger from '~/utils/logger';
 import BaseMap from '~/pages/Reports/components/BaseMap';
 import ClusteredMarkers from './ClusteredMarkers';
-import FMCPropTypes from '~/propTypes';
+import FMCPropTypes from '~/pages/Reports/propTypes';
 
 function toFeature(d) {
   const { geometry, ...properties } = d;
